+++ internal/dataplane/parser/translate_utils.go
@@ -177,6 +177,7 @@ func (p *Parser) generateKongServiceFromBackendRef(
 			},
 			Namespace: route.GetNamespace(),
 			Backends:  backends,
+			Parent:    route,
 		}
 	}