+++ internal/util/objectmeta.go
@@ -1,14 +1,16 @@
 package util
 
 import (
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/apimachinery/pkg/runtime/schema"
+	"sigs.k8s.io/controller-runtime/pkg/client"
 )
 
 // K8sObjectInfo describes a Kubernetes object.
 type K8sObjectInfo struct {
-	Name        string
-	Namespace   string
-	Annotations map[string]string
+	Name             string
+	Namespace        string
+	Annotations      map[string]string
+	GroupVersionKind schema.GroupVersionKind
 }
 
 func deepCopy(m map[string]string) map[string]string {
@@ -19,10 +21,14 @@ func deepCopy(m map[string]string) map[string]string {
 	return result
 }
 
-func FromK8sObject(obj metav1.Object) K8sObjectInfo {
-	return K8sObjectInfo{
+func FromK8sObject(obj client.Object) K8sObjectInfo {
+	ret := K8sObjectInfo{
 		Name:        obj.GetName(),
 		Namespace:   obj.GetNamespace(),
 		Annotations: deepCopy(obj.GetAnnotations()),
 	}
+	if gvk := obj.GetObjectKind().GroupVersionKind(); gvk.String() != "" {
+		ret.GroupVersionKind = gvk
+	}
+	return ret
 }